+++ cashu-lib-common/src/main/java/xyz/tcheeric/cashu/common/util/SecretUtil.java
@@ -8,7 +8,7 @@ import xyz.tcheeric.cashu.common.P2PKSecret;
 import xyz.tcheeric.cashu.common.PublicKey;
 import xyz.tcheeric.cashu.common.RandomStringSecret;
 import xyz.tcheeric.cashu.common.Secret;
-import xyz.tcheeric.cashu.common.VoucherWellKnownSecret;
+import xyz.tcheeric.cashu.common.VoucherSecret;
 import xyz.tcheeric.cashu.common.WellKnownSecret;
 import xyz.tcheeric.cashu.crypto.BDHKEUtils;
 
@@ -160,7 +160,12 @@ public final class SecretUtil<T extends Secret> {
      */
     private static WellKnownSecret createSecret(WellKnownSecret.Kind kind, byte[] data, String nonce) {
         return switch (kind) {
-            case VOUCHER -> new VoucherWellKnownSecret(data, nonce);
+            case VOUCHER -> {
+                VoucherSecret voucher = new VoucherSecret();
+                voucher.setData(data);
+                voucher.setNonce(nonce);
+                yield voucher;
+            }
             case P2PK -> {
                 P2PKSecret p2pk = new P2PKSecret();
                 p2pk.setData(data);